Replace original amplitude with 'Fast' version, as it has full functionality. It is backwards-compatible except for the removed dependence on the beam properties class.
