+++ src/programs/AmplitudeAnalysis/twopi_plotter/twopi_plotter.cc
@@ -21,7 +21,6 @@
 #include "AMPTOOLS_DATAIO/TwoPiPlotGenerator.h"
 #include "AMPTOOLS_DATAIO/ROOTDataReader.h"
 #include "AMPTOOLS_AMPS/TwoPiAngles.h"
-#include "AMPTOOLS_AMPS/TwoPiAnglesFast.h"
 #include "AMPTOOLS_AMPS/BreitWigner.h"
 
 typedef TwoPiPlotGenerator PlotGen;
@@ -29,7 +28,6 @@ typedef TwoPiPlotGenerator PlotGen;
 void atiSetup(){
   
   AmpToolsInterface::registerAmplitude( TwoPiAngles() );
-  AmpToolsInterface::registerAmplitude( TwoPiAnglesFast() );
   AmpToolsInterface::registerAmplitude( BreitWigner() );
   AmpToolsInterface::registerDataReader( ROOTDataReader() );
 }
